requirements.txt
numpy==1.16.2
Pillow==6.0.0
scikit-image==0.15.0
